requirements.txt
tensorflow-gpu==1.4
OpenNMT-tf
tqdm